requirements.txt
pillow==9.5.0
aiogram==2.25.1
requests==2.31.0
python-dotenv==1.0.0
